Add 2 node graph with two self loops and AI properties
This commit contains a two node graph each with a self loop. The graph
has AI properties, specifically a label, a split, and a 2 length feature
vector.

A README containing generation instructions is included as well; because
csv-import doesn't support feature vectors, there is a manual component
to generation. A migrate script exists as well to move from storage
version to storage version as necessary.