Change two_self_loops_ai input to use a generate script
Since I added the infrastructure to use generate scripts for AI inputs,
it only makes sense to change the way two_self_loops_ai is generated as
well.

This commit removes the migrate/add_features scripts and replaces them
with a single generate script. Version 8 has been regenerated with this
script to make sure it works as well.